Updated Card and Frame
Reworked Frame.java to use private static variable to control the deck rather than outsourcing to another class. Also added the deal method as well as mechanics relating to the discard pile

Fixed card to include a joker, where if the Card constructor is called with no parameters, a joker is initialized instead. This joker equals all other cards, has a value of 0, and a suit of "joker," and the toString() method simply returns "joker".